+++ esp-radio/src/wifi/ap.rs
@@ -6,7 +6,7 @@ use core::fmt;
 use enumset::EnumSet;
 use procmacros::BuilderLite;
 
-use super::{AuthMethod, Country, Protocol, SecondaryChannel};
+use super::{AuthenticationMethod, Country, Protocol, SecondaryChannel};
 use crate::{WifiError, sys::include::wifi_ap_record_t};
 
 /// Information about a detected Wi-Fi access point.
@@ -27,7 +27,7 @@ pub struct AccessPointInfo {
     /// The signal strength of the access point (RSSI).
     pub signal_strength: i8,
     /// The authentication method used by the access point.
-    pub auth_method: Option<AuthMethod>,
+    pub auth_method: Option<AuthenticationMethod>,
     /// The country information of the access point (if available from beacon frames).
     pub country: Option<Country>,
 }
@@ -47,7 +47,7 @@ pub struct AccessPointConfig {
     /// The set of protocols supported by the access point.
     pub(crate) protocols: EnumSet<Protocol>,
     /// The authentication method to be used by the access point.
-    pub(crate) auth_method: AuthMethod,
+    pub(crate) auth_method: AuthenticationMethod,
     /// The password for securing the access point (if applicable).
     #[builder_lite(reference)]
     pub(crate) password: String,
@@ -86,7 +86,7 @@ impl Default for AccessPointConfig {
             channel: 1,
             secondary_channel: None,
             protocols: (Protocol::P802D11B | Protocol::P802D11BG | Protocol::P802D11BGN),
-            auth_method: AuthMethod::None,
+            auth_method: AuthenticationMethod::None,
             password: String::new(),
             max_connections: 255,
             dtim_period: 2,
@@ -157,7 +157,7 @@ pub(crate) fn convert_ap_info(record: &wifi_ap_record_t) -> AccessPointInfo {
         channel: record.primary,
         secondary_channel: SecondaryChannel::from_raw(record.second),
         signal_strength: record.rssi,
-        auth_method: Some(AuthMethod::from_raw(record.authmode)),
+        auth_method: Some(AuthenticationMethod::from_raw(record.authmode)),
         country: Country::try_from_c(&record.country),
     }
 }